Added basic great move detection
I have added an algorithm that can detect 'great' moves. These moves are characterised by being the only move in this position that does not remove a relevant part of the player's advantage. There are also some other smaller checks that take place in order to prevent giving 'great' classifications to obvious moves where any other move would be 'dumb', like when a free queen is available. Some classifications of forced mate lines were a bit harsh and blunders were being given if a move that significantly lengthened a forced mate line was played; this has now been made more lenient and a 'mistake' is the worst possible classification in this situation now. Smaller addition is that there is now a small message to confirm that an analysis savefile has been made.