Patched some rendering and classification bugs
Fixed a bug where the program would crash if you tried to use the right arrow key at the end of the game. There was also a bug where sometimes a move that checked the opponent king in such a way that the checking piece was next to the king would be given brilliant even if such piece was defended. I think that I have fixed this bug now but I need to do further testing with other sample games to verify this.