Rewrote brilliant/great move algorithm
The brilliant move classification algorithm that I had previously wrote had lots of bugs in it; brilliant would sometimes be given to rather obvious or not-so-special moves. I have now completely rewritten the brilliant move algorithm and most of the great move algorithm too, and it works much better. Although, there are still a couple things I need to refine with it. I need to slightly tweak the criteria for getting a brilliant move and write logic to prevent easy discovered attack sacrifices being given brilliant.